+++ Service/VideoGrabberWrapper.h
@@ -17,6 +17,6 @@
 - (bool)isFrameNew;
 - (void)drawMirror;
 - (void)draw;
-- (int)getWidth;
-- (int)getHeight;
+- (float)getWidth;
+- (float)getHeight;
 @end
